+++ examples/sampling/plot_sampling_data.py
@@ -7,25 +7,24 @@ within the model domain.
 """
 ###############################################################################
 # First, load the required modules.
-import astropy.constants as const
 import astropy.units as u
 import matplotlib.pyplot as plt
 import numpy as np
 
 from psipy.data import sample_data
-from psipy.model import MASOutput
+from psipy.model import PLUTOOutput
 
 ###############################################################################
 # Load a set of MAS output files, and get the number density variable from the
 # model run. This example works with PLUTO data too. Uncomment the comment
 # block below (and comment out the MAS block!)
 """
-from psipy.model import PLUTOOutput
-pluto_path = sample_data.plut_sample_data()
-model = PLUTOOutput(mas_path)
-"""
+from psipy.model import MASOutput
 mas_path = sample_data.mas_sample_data()
 model = MASOutput(mas_path)
+"""
+pluto_path = sample_data.pluto_sample_data()
+model = PLUTOOutput(pluto_path)
 rho = model["rho"]
 
 ###############################################################################
@@ -40,7 +39,7 @@ rho = model["rho"]
 fig, ax = plt.subplots()
 
 npoints = 1000
-r = 50 * np.ones(npoints) * const.R_sun
+r = 50 * np.ones(npoints) * u.R_sun
 lon = np.linspace(0, 360, npoints) * u.deg
 
 for latitude in [0, 1] * u.deg:
@@ -55,4 +54,5 @@ ax.set_ylim(bottom=0)
 ax.set_xlabel("Longitude (deg)")
 ax.set_ylabel(r"$\rho$ (cm$^{-3}$)")
 ax.set_xticks([0, 90, 180, 270, 360])
+ax.set_title(f"r={r[0].to_value(u.R_sun):.02f} R_sun")
 plt.show()