+++ psipy/tracing/tests/test_tracing.py
@@ -23,7 +23,7 @@ def test_tracer():
 
     tracer = FortranTracer()
 
-    r = 40
+    r = 40 * u.R_sun
     lat = 0 * u.deg
     lon = 0 * u.deg
 
@@ -47,7 +47,7 @@ def test_fline_io(mas_model, tmpdir):
     model = MASOutput(mas_path)
     tracer = FortranTracer()
 
-    r = 40
+    r = 40 * u.R_sun
     lat = 0 * u.deg
     lon = 0 * u.deg